+++ tests/algo/test_sort_all_parameters.cpp
@@ -36,7 +36,7 @@ void test(stxxl::uint64 data_mem, unsigned memory_to_use)
     typedef stxxl::vector<T, 2, stxxl::lru_pager<8>, block_size, alloc_strategy_type> vector_type;
     vector_type v(records_to_sort);
 
-    unsigned ndisks = stxxl::config::get_instance()->disks_number();
+    size_t ndisks = stxxl::config::get_instance()->disks_number();
     STXXL_MSG("Sorting " << records_to_sort << " records of size " << sizeof(T));
     STXXL_MSG("Total volume " << (records_to_sort * sizeof(T)) / MB << " MiB");
     STXXL_MSG("Using " << memory_to_use / MB << " MiB");
